Abstract tab logic to its own component
